+++ src/lib/components/overlays/popover/PopoverGroup.svelte
@@ -12,12 +12,8 @@
     export const CONTEXT_POPOVER_FOCUS_TARGET =
         make_scoped_store<PROPERTY_REFERENCE_TARGET>("popover-focus-target");
 
-    export const CONTEXT_POPOVER_DISMISSIBLE = make_scoped_store<boolean>("popover-dismissible");
-
     export const CONTEXT_POPOVER_LOADING =
         make_scoped_store<PROPERTY_BEHAVIOR_LOADING_LAZY>("popover-loading");
-
-    export const CONTEXT_POPOVER_ONCE = make_scoped_store<boolean>("popover-once");
 
     export const CONTEXT_POPOVER_VARIATION =
         make_scoped_store<PROPERTY_VARIATION_POPOVER>("popover-variation");
@@ -36,9 +32,7 @@
 
         focus_target?: PROPERTY_REFERENCE_TARGET;
 
-        dismissible?: boolean;
         loading?: PROPERTY_BEHAVIOR_LOADING_LAZY;
-        once?: boolean;
 
         variation?: PROPERTY_VARIATION_POPOVER;
     };
@@ -54,9 +48,7 @@
 
     export let focus_target: $$Props["focus_target"] = undefined;
 
-    export let dismissible: $$Props["dismissible"] = undefined;
     export let loading: $$Props["loading"] = undefined;
-    export let once: $$Props["once"] = undefined;
 
     export let variation: $$Props["variation"] = undefined;
 
@@ -65,9 +57,7 @@
 
     const _popover_focus_target = CONTEXT_POPOVER_FOCUS_TARGET.create(focus_target);
 
-    const _popover_dismissible = CONTEXT_POPOVER_DISMISSIBLE.create(dismissible);
     const _popover_loading = CONTEXT_POPOVER_LOADING.create(loading);
-    const _popover_once = CONTEXT_POPOVER_ONCE.create(once);
 
     const _popover_variation = CONTEXT_POPOVER_VARIATION.create(variation);
 
@@ -90,9 +80,7 @@
     $: if (_popover_focus_target)
         $_popover_focus_target = focus_target as PROPERTY_REFERENCE_TARGET;
 
-    $: if (_popover_dismissible) $_popover_dismissible = dismissible as boolean;
     $: if (_popover_loading) $_popover_loading = loading as PROPERTY_BEHAVIOR_LOADING_LAZY;
-    $: if (_popover_once) $_popover_once = once as boolean;
 
     $: if (_popover_variation) $_popover_variation = variation as PROPERTY_VARIATION_POPOVER;