+++ src/lib/components/overlays/popover/PopoverSection.svelte
@@ -26,20 +26,15 @@
     import {TOKENS_TRANSITION_NAMES} from "../../../types/transitions";
 
     import {auto_focus} from "../../../actions/auto_focus";
-    import {click_inside} from "../../../actions/click_inside";
-    import {click_outside} from "../../../actions/click_outside";
     import type {IForwardedActions} from "../../../actions/forward_actions";
     import {forward_actions} from "../../../actions/forward_actions";
-    import {lost_focus} from "../../../actions/lost_focus";
 
     import {map_data_attributes, map_global_attributes} from "../../../util/attributes";
 
     import {
-        CONTEXT_POPOVER_DISMISSIBLE,
         CONTEXT_POPOVER_FOCUS_TARGET,
         CONTEXT_POPOVER_ID,
         CONTEXT_POPOVER_LOADING,
-        CONTEXT_POPOVER_ONCE,
         CONTEXT_POPOVER_STATE,
         CONTEXT_POPOVER_VARIATION,
     } from "./PopoverGroup.svelte";
@@ -97,33 +92,20 @@
 
     const _popover_focus_target = CONTEXT_POPOVER_FOCUS_TARGET.get();
 
-    const _popover_dismissible = CONTEXT_POPOVER_DISMISSIBLE.get();
     const _popover_loading = CONTEXT_POPOVER_LOADING.get();
-    const _popover_once = CONTEXT_POPOVER_ONCE.get();
 
     const _popover_variation = CONTEXT_POPOVER_VARIATION.get();
 
-    function on_dismiss(): void {
-        if (_popover_dismissible && $_popover_dismissible && _popover_state && $_popover_state) {
-            $_popover_state = false;
-        }
-    }
-
-    function on_once(): void {
-        if (_popover_once && $_popover_once && _popover_state && $_popover_state) {
-            $_popover_state = false;
-        }
-    }
-
-    $: _animatable =
+    $: _can_animate =
         (_popover_id && $_popover_id) ||
-        (_popover_variation && $_popover_variation === TOKENS_VARIATION_POPOVER.tooltip);
-    $: _enabled =
+        (_popover_variation && $_popover_variation !== TOKENS_VARIATION_POPOVER.popover);
+    $: _can_auto_focus =
         _popover_state &&
         $_popover_state &&
         _popover_variation &&
         $_popover_variation === TOKENS_VARIATION_POPOVER.popover;
-    $: _label = _popover_id ? `label[for="${$_popover_id}"]` : undefined;
+
+    // TODO: support auto focus in `variation=control`
 </script>
 
 <section
@@ -131,11 +113,11 @@
     {...map_global_attributes($$restProps)}
     class="popover--section {_popover_id ? 'transition' : ''} {_class}"
     {...map_data_attributes({
-        animation: _animatable ? animation ?? TOKENS_TRANSITION_NAMES.clip : undefined,
+        animation: _can_animate ? animation ?? TOKENS_TRANSITION_NAMES.clip : undefined,
         "alignment-x": alignment_x,
         "alignment-y": alignment_y,
-        behavior: _animatable ? "explicit" : "",
-        direction: _animatable
+        behavior: _can_animate ? "explicit" : "",
+        direction: _can_animate
             ? DIRECTIONS_LOOKUP[placement ?? TOKENS_DIRECTIONS.bottom]
             : undefined,
         placement,
@@ -143,21 +125,8 @@
         "spacing-x": spacing_x,
         "spacing-y": spacing_y,
     })}
-    use:click_inside={{
-        ignore: _label,
-        on_click_inside: on_once,
-    }}
-    use:click_outside={{
-        ignore: _label,
-        on_click_outside: on_dismiss,
-    }}
-    use:lost_focus={{
-        enabled: _enabled,
-        ignore: _label,
-        on_lost_focus: on_dismiss,
-    }}
     use:auto_focus={{
-        enabled: _enabled,
+        enabled: _can_auto_focus,
         target: _popover_focus_target ? $_popover_focus_target : null,
     }}
     use:forward_actions={{actions}}