+++ src/lib/components/widgets/dataselect/index.ts
@@ -1,0 +1,1 @@
+export {default as DataSelect} from "./DataSelect.svelte";