+++ src/lib/components/widgets/inlay/index.ts
@@ -1,0 +1,1 @@
+export {default as Inlay} from "./Inlay.svelte";